+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.1.6 - 2021-07-26
+- Fixed support for nested arrays
+- Improved support for complex types
+- Improved test coverage
+
 ### 0.1.5 - 2021-06-27
 - Removed `flow-js-testing` dependency
 - Refactored code generation for all types of interactions
@@ -6,10 +11,10 @@
 - Updated `flow-generate` CLI. Added GitHub support and flags.
 
 ### 0.1.4 - 2021-06-11
-- Update regexp matchers to properly catch script and transaction arguments
+- Updated regexp matchers to properly catch script and transaction arguments
 
 ### 0.1.3 - 2021-06-10
-- Export mapArgument, mapArguments and mapValuesToCode methods.
+- Exported mapArgument, mapArguments and mapValuesToCode methods.
 
 ### 0.1.2 -- 2021-06-02
 - Regenerated templates to remove unused import, which would trigger linter.